+++ silhouette/test/com/mohiva/play/silhouette/impl/providers/oauth2/state/CookieStateSpec.scala
@@ -48,7 +48,7 @@ class CookieStateSpec extends PlaySpecification with Mockito with JsonMatchers {
       val dateTime = new DateTime(2014, 8, 8, 0, 0, 0)
       val decoded = Base64.decode(state.copy(expirationDate = dateTime).serialize)
 
-      decoded must /("expirationDate" -> dateTime.getMillis)
+      decoded must /("expirationDate" -> dateTime.getMillis.toDouble)
       decoded must /("value" -> "value")
     }
   }
